employee_management/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
requests
pillow
django-cors-headers
gunicorn
